requirements.txt
grpcio
tensorflow==1.0.0
